third_party/WebKit/LayoutTests/svg/custom/keySplines-parsing-error-expected.txt

CONSOLE ERROR: line 4: Error: <animate> attribute keySplines: Invalid value, ";;".
CONSOLE ERROR: line 5: Error: <animate> attribute keySplines: Invalid value, "0 ,0  1 , 1  ;;".
Tests parsing of keySplines attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
